Update GoogleTest to its latest version (current master).
The new version can be built faster:

    Original version:

        real    0m11.028s
        user    0m10.241s
        sys     0m0.782s

    Latest version:

        real    0m5.178s
        user    0m4.753s
        sys     0m0.425s

As a related change, this commit adds missing dependencies between the
gtest_main, gmock, and gmock_main libraries. In the original version,
libgmock.a contained libgtest.a and libgmock_main.a contained libgmock.a. Now,
libgmock.a is without libgtest.a and libgmock_main.a is without libgmock.a.
Thus, if one wants to link gmock_main, he or she has to also link gmock and
gtest. This has been reflected in the dependencies.