scripts: Fix test discovery in retdec-tests-runner.sh on macOS (#238).
On macOS, find does not support the '-executable' parameter (#238). Therefore,
on macOS, we have to use '-perm +111'. To explain, + means "any of these bits"
and 111 is the octal representation for the executable bit on owner, group, and
other. Unfortunately, we cannot use '-perm +111' on all systems because find on
Linux/MSYS2 does not support +. It supports only /, but this is not supported
by find on macOS... Hence, we need an if.