fileformat: change getByteLength and getNibbleLength
Previously, those functions returned 0 if architecture was
unknown but this broke a lot of functions using these
functions if architecture that was not supported by
decompiler was encountered. Now functions will return
values 4 and 8. This will break with some excotic
architectures that use other size of byte but those wouldn't
work anyway.