fileformat: Verify digital signature after we have a signer/counter-signer (#87).
In #87, a fileinfo crash is reported when verifying the digital signature of
attached PE files. What all the attached files have in common is that we are
unable to find a signer or counter-signer for them and p7->length is 0. As the
following comment in pe_format.cpp suggests, there is no point of continuing in
such a case:

    // If we have no signer and countersigner, there must be something really bad
    if(!signerCert && !counterSignerCert)
    {
        BIO_free(bio);
        return;
    }

Thus, move the signature verification AFTER the check that we have found a
signer or a counter-signer. This fixes the signature-verifying crashes for all
the files attached to #87.