scripts: Fix the used sed commands at the end of retdec-decompiler.sh to work on macOS (#257).
The original commands failed with the following error on macOS:

    sed: 1: "/private/tmp/test.c": undefined label 'mp/test.c'
    sed: 1: "/private/tmp/test.c": undefined label 'mp/test.c'

The reason why the commands failed is that when using the -i parameter of sed
on macOS, you need to specify an extension of the backup file. If you do not
want to create a backup file, you have to pass the empty string. On Linux,
using just -i without any extension suffices.

The new sed command should work on Linux, macOS, and Windows, without a need to
install gnu-sed on macOS.