scripts: Prevent retdec-decompiler.sh from creating output.c-- files.
This fixes the used sed command from commit
7b5289acda02cfb8fbbf3a618c5a57c3f9c5a2a2, which caused "output.c--" files to be
created when running the decompiler on macOS.

Apparently, there is no way of getting sed -i to work consistently on both
MacOS and Linux. Indeed, this works with GNU sed, but not on macOS:

    sed -i 's/foo/bar/' FILE
    sed -i'' 's/foo/bar/' FILE

This works on macOS, but not with GNU sed:

    sed -i '' 's/foo/bar/' FILE

So, store the output from sed into a temporary file and rename that file
afterwards to get consistent behavior between macOS and Linux.