scripts: Use gtimeout instead of timeout when available (#260).
Before this commit, retdec-archive-decompiler.sh on macOS failed with the
following error:

    $ retdec-archive-decompiler.sh
    error: The `timeout` command is required but it is not available. Aborting.

The reason for this is that on macOS, timeout from GNU coreutils is installed
as gtimeout (unless you have installed coreutils with `brew install coreutils
--with-default-names`).

The present commit updates the script so it uses gtimeout when available (we
already do the same for e.g. greadlink and gha256sum).