scripts: Add new parameters to retdec-decompiler.sh: --max-memory and --no-memory-limit.
The first parameter can be used to set the maximal memory limit of fileinfo,
bin2llvmir, and llvmir2hll.

The second parameter can be used to skip the newly added limit of half of
system RAM in fileinfo, bin2llvmir, llvmir2hll. By default, we want to limit
the memory of these tools into half of system RAM to prevent potential black
screens on Windows.