llvm: Disable forced reconfiguration to speed up RetDec build and installation.
This was a relict from our old build system, where we built LLVM as an external
project from a subdirectory and needed to detect changes in its source files.
Now, we build LLVM as an external project from a GitHub URL, so there is no
longer a need to re-configure LLVM after it has been built (its sources will
never change).

This speeds up the rebuild and installation of RetDec. For example, on my Linux
machine, a rebuild without changes has speeded up from

	real    0m1.223s
	user    0m2.956s
	sys     0m1.163s

to

	real    0m0.527s
	user    0m1.903s
	sys     0m0.716s

On Windows, the difference will be even bigger. This lowering of rebuild times
is very handy during development.