scripts: Add a missing extension to files generated by retdec-archive-decompiler.sh (#261).
The retdec-archive-decompiler.sh script did not append .c at the end of the
decompiled files from the input archive (see #261). This commit fixes that.

Since there is currently no way of selecting the target language when running
retdec-archive-decompiler.sh, we can just append '.c' to the output files as C
is the default target language.