unpacker: Improve copying of PE section names to suppress GCC 8 warning (#325).
GCC 8 added a new warning called `-Wstringop-truncation`. From
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=82944:

    The -Wstringop-truncation warning added in GCC 8.0 via r254630 for bug
    81117 is specifically intended to highlight likely unintended uses of the
    strncpy function that truncate the terminating NUL charcter from the source
    string. An example of such a misuse given in the request is the following:

        char buf[2];

        void test(const char* str) {
            strncpy(buf, str, strlen(str));
        }

Since section names in PE files have a fixed size of 8 bytes, it is fine to
have a name without a terminating null byte. Thus, replace strncpy() with
memcpy() to suppress the warning and make the code more readable (we are
actually copying bytes, not characters).