scripts/type_extractor: replace python scripts with the original bash scripts.
Python scripts were not working. They were also relying on shell utilities (e.g. grep, find). Users do not need to execute these, so these do not need to be rewritten to python to get rid of MSYS on Windows. We can deal with this later - they need to be completely rewritten in Python, not converted from Bash.