Change SCC's to be FuncVectorSet instead of FuncSetSet.
Previously, the set of sccs was sorted by overall function pointer order.
This destroyed the topological ordering of SCCs that tarjan creates.

Also, insertion of each scc possibly compared it against log N sets,
where each comparison is O(N) (The definition of operator< on sets
performs lexicographic compare on the members of each set).
This did not do well in the face of large sccs.