bin2llvmir/filter: check validity of return value
When parameter filtration finds stack parameters in function
definition but not in function call it crieates dummy
call argument that is passed to the called function. This is
because we do not want to lost argumetns that are known
from function definition. This dummy call argument
is during filtration represented as nullptr so it won't
affect non-nullptr argument values of other calls.

This is why it is needed to check definition of stack
variable offset as some parts of filter or
new derived filters may not provide valid value for filtration.