cpdetect: Fix a signedness warning in PeHeuristics::skip_NOP_JMP8_JMP32().
This fixes the following GCC warning on Linux:

    src/cpdetect/compiler_detector/heuristics/pe_heuristics.cpp:370:65:
        warning: comparison of integer expressions of different signedness:
        ‘int32_t’ {aka ‘int’} and ‘unsigned int’ [-Wsign-compare]
    if ((codePtr + 5) > codeEnd || getInt32Unaligned(codePtr+1) == 0x80000000)
                                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~

To elaborate, 0x80000000 as int32_t is -2147483648, which is the minimal
integer representable in 32 bits. The compiler recognizes 0x80000000
(2147483648) as an unsigned integer as this number cannot be represented as a
signed 32b integer. Since getInt32Unaligned() returns int32_t, we are comparing
values having different signedness. Fix this by turning 0x80000000 into a
signed integer.