capstone2llvmir: arm64: store to FP register requires different ct
Currently there is not a good implementation/design for `storeOp`
methods, that store LLVM Values into other operands.
As there may occur errors f.e. on the side of capstone, we have to
make sure that operands are converted correctly.

This commit brings only temprorary solution for arm64
as better refactorization of store operands method is required.
This hovewer affects all architectures and design shuld be better
rethought.