capstone2llvmir/capstone2llvmir_impl: mark method internal
Provides marking of the method `checkTypeConvertion` as internal
by giving it underscore prefix (`_checkTypeConvertion`). This
will indicate that this method is ment only to be used as a specific
helper method and not to be used with other methods.

Also provides documentation for this method as the name of the
method may suggest different functionality.