Try to fix issue #633 (#672)
* capston2llvmir_impl: provide better support for operands loading

* capstone2llvmir: arm64: store to FP register requires different ct

Currently there is not a good implementation/design for `storeOp`
methods, that store LLVM Values into other operands.
As there may occur errors f.e. on the side of capstone, we have to
make sure that operands are converted correctly.

This commit brings only temprorary solution for arm64
as better refactorization of store operands method is required.
This hovewer affects all architectures and design shuld be better
rethought.

* capston2llvmir: arm: store to FP reg requires different ct

Currently there is not a good implementation/design for `storeOp`
methods, that store LLVM Values into other operands.
As there may occur errors f.e. on the side of capstone, we have to
make sure that operands are converted correctly.

This commit brings only temprorary solution for arm
as better refactorization of store operands method is required.
This hovewer affects all architectures and design shuld be better
rethought.

* capston2llvmir: provide documentation for internal method

* capstone2llvmir/capstone2llvmir_impl.h: fix doxygen

* capstone2llvmir/capston2llvmir_impl: fix type in 825efcab

The commit 825efcab brought unwanted change of ranaming `loadOp` method
from `loadOpUnary` to `loadOpsUnary` which is not consistent
with naming of other methods. This commit brings fix for this
change and therefore the old name of the method.

* capstone2llvmir/capstone2llvmir_impl: mark method internal

Provides marking of the method `checkTypeConvertion` as internal
by giving it underscore prefix (`_checkTypeConvertion`). This
will indicate that this method is ment only to be used as a specific
helper method and not to be used with other methods.

Also provides documentation for this method as the name of the
method may suggest different functionality.

* tests/capstone2llvmir: add binary emulation mode.

Sometimes, Keystone refuses to assembly the given instruction. Then, we can not translate and emulate it. This adds a possibility to directly emulate binary data - skipping the assembly process.

* CHANGELOG.md: fix #633, add entry for #633 and #672.

* tests/capstone2llvmir/mips_tests: add unit test for #633.