bin2llvmir/param_return: provide user defined function signatures
This reverts commit 5cbad885c3186055263d599b86263ef1fa73ae77.

This commit bings and fixes changes brought by original commit.
Original commit tried to provide option to load custom function
signature from json configuration file.

As there was confusion in the type of functions (userDefined,
staticallyLinked, etc.) the heurisics did not work as expected
and caused some tests to fail. This commit brings changes
that fix this issue and provide improvements in the output.