common/function: new function type DECOMPILER_DEFINED
Before this commit function type might be one of the following:
  - USER_DEFINED
  - STATICALLY_LINKED
  - DYNAMICALLY_LINKED
  - SYSCALL
  - IDIOM

The default function type that was generated for every function was
USER_DEFINED. This is not right as we have to distinguish functions that
were provided by user through some kind of front-end and functions that
were automatically detected by decompiler. This is why this commit
brings new link type - DECOMPILER_DEFINED.

A reason to distinguish these two link types is for example in parameter
analysis. When user specifies that the return type is integer, the decompiler
should not continue to do heuristics to detect return type. If an user
wants a function to be of some type they should get the function with the type
on output.

As USER_DEFINED now requires interaction of an user, the DECOMPILER_DEFINED
should become the default link type.