debuformat/dwarf: use only valid name of a object
This commit provides fix for an unexpected asserions in some cases. This
problem was caused by not checking the output of a function that was
used to fetch name of a object. This name could have been an empty
string. In such case the constructor of a common::Object would assert in
debug mode.