param_return: Fix MS x64 ABI detecion (#707).
This commit fixes detection of Microsoft x64 ABI. The original detection
was based on Microsoft compiler AND file format. This commit generalizes the
condtion to only the file format OR Microsoft compiler.

So MsX64 convention is used when the file format is PE or the binary was
created by a Microsoft based compiler.