deps/yara: fix build on macOS
Fixes bug on macOS that caused building error in some cases.
The `xcrun --show-sdk-path` would append new line on the end of the
MACOS_SDK_PATH variable. This would cause error when multiple arguments
are passed to the CFLAGS variable of the configure command.

The fix is to provide OUTPUT_STRIP_TRAILING_WHITESPACE option to the
`executre_process` command that would leave out the new line character
from the output.