deps/yara: fix possible build error
Fixes error when an user had set more than one CMAKE_C_FLAGS options.
In this case it had caused build to fail. The reason for this error to happen
was that the CMAKE_C_FLAGS was passed to the `./configure` script as

`$ ./configure CFLAGS=${CMAKE_C_FLAGS}`.

In CMakeLists.txt this was caused by CMAKE_C_FLAGS variable being list.
Passing list to configure would cause the list to be devided and interpreted
as more options and `./configure` would not recognize all of them.

The solution is to set the CMAKE_C_FLAGS variable to be string. In this case
the command will be `./configure CFLAGS="${CMAKE_C_FLAGS}"` (Notice the
doubleqoutes).