Disable passing processor count in CMakeLists.txt
Provides disabling of processor count in CMakeLists.txt.
POSIX/Windows Makefiles use jobserver method when running in parallel.
This means that first invokation of make creates a jobserver
and each command that can communicate with jobserver is provided
with access to do job when a active job is available. Such
command that can access jobserver is command make.

Specification of `-j` in the make command in CMakeLists.txt
results in disabling of the jobserver. This should mean
that each make command with option `-j` creates new jobserver
with new job pool.

More about jobservers:
https://www.gnu.org/software/make/manual/html_node/Job-Slots.html