Prevent false detections of VMProtect.
By rigorously checking the imported API and the presence of the string
in only executable sections of the file we ensure that only truly packed
files are detected. Otherwise a false positive detection could occur when
a file was doubly packed and the byte pattern was stored in unmodified
form in the data section, or when a packed file was stored in resources
or overlay.