param_return: check possible error in stack analysis
There might be an error in stack analysis. In suuch cases we
would not like to trust information gathered from analysis
of function body.

This should be changed in future by incorporating refactorization
of data_entries module design. Right now parameters might be only
stack variables and registers. Not other values -> this will change
as it is required as usage showed.