Allow to configure `yara`'s make tool
This is another approach to remove hardocoded call of `make` to build
`yara`. This approach uses the same `make` by default but introduces
`YARA_MAKE_PROGRAM` option which can be used to specify a path to
`gmake` for example.

Thus, when CMake produces a set of makefile, it reuses
`CMAKE_MAKE_PROGRAM` to respect a global `-jXXX` option.