fix get10Byte on platforms where long double is >80 bits.
for platforms with a non 80-bit "long double" type, we
change get10ByteImpl to reduce to double then use the
language's conversion to long double.

this handles both where long double is smaller and greater
than 80 bits, though only 64 bits of precision are preserved.

this fixes the tests on aarch64 platforms.